+++ cmd/gpscrape/catalog.go
@@ -16,7 +16,7 @@ import (
 	"strings"
 	"time"
 
-	googleplayscraper "github.com/kryuchenko/google-play-scraper"
+	googleplayscraper "github.com/kryuchenko/google-play-scraper/v2"
 )
 
 // The catalog verbs.