+++ examples/app-availability/main.go
@@ -23,7 +23,7 @@ import (
 	"strings"
 	"time"
 
-	googleplayscraper "github.com/kryuchenko/google-play-scraper"
+	googleplayscraper "github.com/kryuchenko/google-play-scraper/v2"
 )
 
 func main() {