+++ examples/yandex-taxi-reviews/main.go
@@ -7,7 +7,7 @@ import (
 	"os"
 	"time"
 
-	"github.com/kryuchenko/google-play-scraper"
+	"github.com/kryuchenko/google-play-scraper/v2"
 )
 
 func main() {